+++ .storybook/main.ts
@@ -1,0 +1,8 @@
+import type { StorybookConfig } from '@storybook/html-vite';
+
+const config: StorybookConfig = {
+  stories: ['../src/stories/**/*.stories.ts'],
+  framework: '@storybook/html-vite',
+};
+
+export default config;